jwt의 주인이 아닌 다른 사람의 데이터를 가져오지 못하도록 수정

+++ ApiDir/BookMark.py
@@ -26,14 +26,13 @@ class BookMarkAdd(Resource):
     @BookMark.expect(parser, BookMark_field)
     @BookMark.expect(BookMark_field)
     def post(self):
-        '''User가 선호하는 웹툰를 저장하는 API\nUser ID와 웹툰 제목을 json의 형태로 전달받아 DB에 저장한다.'''
+        '''User가 선호하는 웹툰를 저장하는 API\nUser ID와 웹툰 제목을 json의 형태로 전달받아 DB에 저장한다. \n uid는 필요없음'''
         # 데이터 파싱
-        UID = request.json.get('UID')
         Title = request.json.get('Title')
 
         try:
             # BookMark에 맞는 형태로 변환 후 session을 열고 저장
-            data = models.BookMark(UID=UID, WebtoonTitle=Title)
+            data = models.BookMark(UID=get_jwt_identity(), WebtoonTitle=Title) # id를 입력 받는 것이 아니라 jwt에서 추출
             db.session.add(data)
             db.session.commit()
             db.session.flush()
@@ -44,14 +43,19 @@ class BookMarkAdd(Resource):
 @BookMark.route('/<UID>')
 class BookMarkList(Resource):
     '''User가 즐겨찾기에 등록한 모든 웹툰들을 쿼리하여 가져오는 api\n\
-        해당 User의 ID와 동일한 UID를 가진 모든 북마크들을 리스트 형태로 받아온다.'''
+        해당 User의 ID와 동일한 UID를 가진 모든 북마크들을 리스트 형태로 받아온다.\n
+        만약 jwt의 주인의 id와 가져올 데이터의 id가 다른 경우 1을 반환한다.'''
     
     @jwt_required() #jwt 검증
     @BookMark.expect(parser)
     def get(self, UID):
-        data = db.session.query(models.BookMark).filter(models.BookMark.UID==UID)
-        data = pd.read_sql(data.statement, data.session.bind)
-        return json.loads(data.to_json(orient='records'))
+        temp_id = get_jwt_identity()
+        if temp_id == UID:
+            data = db.session.query(models.BookMark).filter(models.BookMark.UID==UID)
+            data = pd.read_sql(data.statement, data.session.bind)
+            return json.loads(data.to_json(orient='records'))
+        else:
+            return 1 # 요청한 id와 가져올 데이터의 id가 다른 경우
 
 @BookMark.route('/<UID>/<WebToonTitle>')
 class BookMarkDelete(Resource):
@@ -62,10 +66,14 @@ class BookMarkDelete(Resource):
     @BookMark.expect(parser)
     def delete(self, UID, WebToonTitle):
         
-        try:
-            db.session.query(models.BookMark).filter(models.BookMark.UID==UID, models.BookMark.WebtoonTitle==WebToonTitle).delete()
-            db.session.commit()
-            return 0 # 쿼리 성공 시
-        except:
-            return 1 # 쿼리 실패 시
+        temp_id = get_jwt_identity()
+        if temp_id == UID:
+            try:
+                db.session.query(models.BookMark).filter(models.BookMark.UID==UID, models.BookMark.WebtoonTitle==WebToonTitle).delete()
+                db.session.commit()
+                return 0 # 쿼리 성공 시
+            except:
+                return 1 # 쿼리 실패 시
+        else:
+            return "jwt와 가져올 데이터의 id가 일치하지 않습니다."
